Move categorize, get_segments to dwpc.py
