+++ hetmech/dwpc.py
@@ -204,13 +204,145 @@ def dwpc_long_repeat(graph, metapath, damping=0.5):
     raise NotImplementedError("See PR #59")
 
 
+def categorize(metapath):
+    """
+    Returns the classification of a given metapath as one of
+    a set of metapath types which we approach differently.
+    Parameters
+    ----------
+    metapath : hetio.hetnet.MetaPath
+    Returns
+    -------
+    classification : string
+        One of ['no_repeats', 'disjoint', 'short_repeat',
+                'long_repeat', 'BAAB', 'BABA', 'other']
+    Examples
+    --------
+    GbCtDlA -> 'no_repeats'
+    GiGiG   -> 'short_repeat'
+    GiGiGcG -> 'long_repeat'
+    GiGcGiG -> 'long_repeat'
+    GiGbCrC -> 'disjoint'
+    GbCbGbC -> 'BABA'
+    GbCrCbG -> 'BAAB'
+    DaGiGbCrC -> 'disjoint'
+    GiGaDpCrC -> 'disjoint'
+    GiGbCrCpDrD -> 'disjoint'
+    GbCpDaGbCpD -> NotImplementedError
+    GbCrCrCrCrCbG -> NotImplementedError
+    """
+    metanodes = list(metapath.get_nodes())
+    repeated_nodes = {v for i, v in enumerate(metanodes) if
+                      v in metanodes[i + 1:]}
+
+    if not repeated_nodes:
+        return 'no_repeats'
+
+    repeats_only = [node for node in metanodes if node in repeated_nodes]
+
+    # Group neighbors if they are the same
+    grouped = [list(v) for k, v in itertools.groupby(repeats_only)]
+
+    # Handle multiple disjoint repeats, any number, ie. AA,BB,CC,DD,...
+    if len(grouped) == len(repeated_nodes):
+        # Identify if there is only one metanode
+        if len(set(metanodes)) == 1:
+            if len(metanodes) < 4:
+                return 'short_repeat'
+            else:
+                return 'long_repeat'
+
+        return 'disjoint'
+
+    # Group [A, BB, A] or [A, B, A, B] into one
+    if len(repeats_only) - len(grouped) <= 1:
+        grouped = [repeats_only]
+
+    # Categorize the reformatted metapath
+    if len(grouped) == 1 and len(grouped[0]) == 4:
+        if grouped[0][0] == grouped[0][-1]:
+            return 'BAAB'
+        else:
+            return 'BABA'
+    else:
+        # Multi-repeats that aren't disjoint, eg. ABCBAC
+        if len(repeated_nodes) > 2:
+            raise NotImplementedError(
+                "Only two overlapping repeats currently supported")
+
+        if len(metanodes) > 5:
+            raise NotImplementedError(
+                "Complex metapaths of length > 4 are not yet supported")
+        return 'other'
+
+
 def get_segments(metagraph, metapath):
-    """Should categorize things into more than just the five categories
-    in PR # 60. We want to segment the metapath into long-repeats, short-
-    repeats, BABA, (which can not at the moment include other intermediates),
-    BAAB (which can have intermediates as long as the whole thing is
-    symmetrical), and other, non-segment-able regions."""
-    raise NotImplementedError("Will integrate PR #60")
+    """
+    Split a metapath into segments of recognized groups and non-repeated
+    nodes. Groups include BAAB, BABA, disjoint short- and long-repeats.
+    Returns an error for categorization 'other'.
+
+    Parameters
+    ----------
+    metagraph : hetio.hetnet.MetaGraph
+    metapath : hetio.hetnet.Metapath
+
+    Returns
+    -------
+    list
+        list of metapaths. If the metapath is not segmentable or is already
+        fully simplified (eg. GaDaGaD), then the list will have only one
+        element.
+
+    Examples
+    --------
+    'CbGaDaGaD' -> ['CbG', 'GaDaGaD']
+    'GbCpDaGaD' -> ['GbCpDaGaD']
+    'CrCbGiGaDrD' -> ['CrCbG', 'GiGaD', 'DrD']
+    """
+
+    def add_head_tail(metapath, indices):
+        # handle non-duplicated on the front
+        if indices[0][0] != 0:
+            indices = [[0, indices[0][0]]] + indices
+        # handle non-duplicated on the end
+        if indices[-1][-1] != len(metapath):
+            indices = indices + [[indices[-1][-1], len(metapath)]]
+        return indices
+
+    category = categorize(metapath)
+    metanodes = metapath.get_nodes()
+    freq = collections.Counter(metanodes)
+    repeated_nodes = {i for i in freq.keys() if freq[i] > 1}
+
+    if category == 'other':
+        raise NotImplementedError("Incompatible metapath")
+
+    elif category in ('disjoint', 'short_repeat', 'long_repeat'):
+        indices = sorted([[metanodes.index(i), len(metapath) - list(
+            reversed(metanodes)).index(i)] for i in repeated_nodes])
+        indices = add_head_tail(metapath, indices)
+        # handle middle cases with non-repeated nodes between disjoint regions
+        # Eg. [[0,2], [3,4]] -> [[0,2],[2,3],[3,4]]
+        inds = []
+        for i, v in enumerate(indices[:-1]):
+            inds.append(v)
+            if v[-1] != indices[i + 1][0]:
+                inds.append([v[-1], indices[i + 1][0]])
+        indices = inds + [indices[-1]]
+
+    elif category in ('BAAB', 'BABA'):
+        assert len(repeated_nodes) == 2
+        start_of_baba = min([metanodes.index(i) for i in repeated_nodes])
+        end_of_baba = len(metapath) - min(
+            [list(reversed(metanodes)).index(i) for i in repeated_nodes])
+        indices = [[start_of_baba, end_of_baba]]
+        indices = add_head_tail(metapath, indices)
+
+    segments = [metapath[i[0]:i[1]] for i in indices]
+    segments = [i for i in segments if i]
+    segments = [metagraph.get_metapath(metaedges) for metaedges in segments]
+    return segments
 
 
 def dwpc(graph, metapath, damping=0.5):
